# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MathPlatform API",
        "description": "API for MathPlatform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/student/lessons": {
            "get": {
                "tags": [
                    "Student Lessons"
                ],
                "summary": "Get student lessons with learning progress",
                "operationId": "studentLessonsIndex",
                "responses": {
                    "200": {
                        "description": "Lessons retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Sanctum token returned by the admin login.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}